# Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: StudentSocial
  version: "1.0"
  description: Die StudentSocial Website ist eine Single Page App im Frontend und ein REST Webservice im Backend. Dadurch sollen Studenten und andere Objekte in einer Datenbank gespeichert werden.
  contact:
    name: WWI20B4 Gruppe
    email: mai.julian@student.dhbw-karlsruhe.de
    url: "https://www.google.de"
  license:
    name: ABC-Lizenz
    url: "https://www.google.de"
  termsOfService: "https://www.google.de"
servers:
  - url: "http://localhost:3000"
    description: Lokale Entwicklung
paths:
...
  /user:
    get:
      summary: Userliste durchsuchen
      tags:
        - Userliste
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
      operationId: get-user
      description: Userliste durchsuchen
      parameters:
        - schema:
            type: string
          in: query
          name: matrikel_nr
          description: Matrikelnummer suchen
        - schema:
            type: string
          in: query
          name: email
          description: Email suchen
        - schema:
            type: string
          in: query
          name: password
          description: Passwort suchen
    post:
      summary: Neue Adresse anlege
      tags:
        - Userliste
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      operationId: post-user
      description: User erstellen
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: User Parameter über JSON
...
components:
  schemas:
...
    User:
      title: User
      type: object
      description: Ein Eintrag in der User-Tabelle um den Login zu authentifizieren.
      properties:
        _links:
          $ref: "#/components/schemas/HATEOAS_Resource"
        _id:
          type: string
        matrikel_nr:
          type: string
        email:
          type: string
        password:
          type: string
...
    HATEOAS_Resource:
      title: HATEOAS_Resource
      type: object
      description: "HATEOAS-Verlinkungen zu einem einzelnen Datensatz / Adressbucheintrag. Definiert, wie der Eintrag gelesen, aktualisiert, gepatcht oder gelöscht werden kann."
      properties:
        read:
          $ref: "#/components/schemas/HATEOAS_Link"
        update:
          $ref: "#/components/schemas/HATEOAS_Link"
        patch:
          $ref: "#/components/schemas/HATEOAS_Link"
        delete:
          $ref: "#/components/schemas/HATEOAS_Link"
    HATEOAS_Link:
      title: HATEOAS_Link
      type: object
      description: "Eine einzelner HATEOAS-Verlinkung innerhalb des Webservices. Definiert die Adresse, URL und Query Parameter zum Aufruf der jeweiligen Aktion."
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
...